+++ src/aspire/image/__init__.py
@@ -7,4 +7,3 @@ from .image import (
     PolarImage,
     normalize_bg,
 )
-from .xform import CustomNoiseAdder, WhiteNoiseAdder
