+++ src/aspire/source/simulation.py
@@ -6,7 +6,7 @@ from scipy.linalg import eigh, qr
 from sklearn.metrics import mean_squared_error
 
 from aspire.image import Image
-from aspire.image.xform import NoiseAdder, WhiteNoiseAdder
+from aspire.noise import NoiseAdder, WhiteNoiseAdder
 from aspire.source import ImageSource
 from aspire.source.image import _ImageAccessor
 from aspire.utils import (
